moved NytArticle into a scrapers file
