+++ website/frontend/management/commands/scraper.py
@@ -131,53 +131,7 @@ def get_all_article_urls():
         ans = ans.union(map(canonicalize_url, urls))
     return ans
 
-#Parser for NYT articles
-#also used as a base class for other parsers; probably should be split
-class Article(object):
-    url = None
-    title = None
-    date = None
-    body = None
-    real_article = True
-    meta = []
-    SUFFIX = '?pagewanted=all'
-
-    def __init__(self, url):
-        self.url = url
-        self.html = grab_url(url + self.SUFFIX)
-        open('/tmp/moo', 'w').write(self.html)
-        open('/tmp/moo3', 'w').write(url+self.SUFFIX)
-        self._parse(self.html)
-
-    def _parse(self, html):
-        soup = BeautifulSoup(html, convertEntities=BeautifulSoup.HTML_ENTITIES)
-        self.meta = soup.findAll('meta')
-        self.seo_title = soup.find('meta', attrs={'name':'hdl'}).get('content')
-        tmp = soup.find('meta', attrs={'name':'hdl_p'})
-        if tmp and tmp.get('content'):
-            self.title = tmp.get('content')
-        else:
-            self.title = self.seo_title
-        self.date = soup.find('meta', attrs={'name':'dat'}).get('content')
-        self.byline = soup.find('meta', attrs={'name':'byl'}).get('content')
-        p_tags = soup.findAll('p', attrs={'itemprop':'articleBody'})
-        self.body = '\n'.join([p.getText() for p in p_tags])
-        authorids = soup.find('div', attrs={'class':'authorIdentification'})
-        self.authorid = authorids.getText() if authorids else ''
-
-        self.top_correction = '\n'.join(x.getText() for x in
-                                   soup.findAll('nyt_correction_top')) or '\n'
-        self.bottom_correction = '\n'.join(x.getText() for x in
-                                   soup.findAll('nyt_correction_bottom')) or '\n'
-
-    def __unicode__(self):
-        return strip_whitespace(u'\n'.join((self.date, self.title, self.byline,
-                                            self.top_correction, self.body,
-                                            self.authorid,
-                                            self.bottom_correction,)))
-
-
-
+'''
 class PoliticoArticle(Article):
     SUFFIX = ''
 
@@ -307,11 +261,11 @@ class TagesschauArticle(Article):
 
     def __unicode__(self):
         return self.document
-
+'''
 
 feeders = [
-#            ('http://www.nytimes.com/',
-#            lambda url: 'www.nytimes.com/201' in url),
+           ('http://www.nytimes.com/',
+            lambda url: 'www.nytimes.com/201' in url),
            ('http://edition.cnn.com/',
             lambda url: 'edition.cnn.com/201' in url),
 #           ('http://www.politico.com/',
